requirements.txt
Django == 3.0.8
python-twitter == 3.5
psycopg2 == 2.8.5
djangorestframework == 3.11.0